+++ src/logging/default-logger.ts
@@ -170,6 +170,11 @@ export class DefaultLogger implements Logger {
             console.error(`❌${prefix} Process spawn error: ${event.message}`);
             break;
           }
+          case 'cancelled': {
+            // The script never started. We don't really need to log this, it's
+            // fairly noisy. Maybe in a verbose mode.
+            break;
+          }
           case 'unknown-error-thrown': {
             console.error(
               `❌${prefix} Internal error! Please file a bug at https://github.com/google/wireit/issues/new, mention this message, that you encountered it in wireit version ${getWireitVersion()}, and give information about your package.json files.\n    Unknown error thrown: ${String(